windows: Add script that relaunches kmonad on crash
This powershell script launches kmonad, then watches for the error
message "Encountered error in KeySource: Cannot translate from windows
keycode" which means the kmonad process has crashed.

If it sees this, it immediately kills the existing kmonad process and
then launches a new one, again watching for a crash.

It also handles C-c just fine and will kill any existing kmonad
processes before exiting as long as the matching pattern is correctly in
the "finally" block at the end of this file.

Calling the script works like this:

  .\relaunch_kmonad_after_keycode_error.ps1 -binary "C:\Full\Path\To\Your\kmonad_binary.exe" -config "C:\Full\Path\To\Your_Kmonad_Config.kbd"

Both arguments are REQUIRED and both MUST BE surrounded by double
quotes.